requirements.txt
alembic==1.7.5
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==21.4.0
boto3==1.20.28
botocore==1.23.28
click==8.0.3
colorama==0.4.4
contextlib2==21.6.0
factory-boy==3.2.1
Faker==11.1.0
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.29
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
